app/pyproject.toml
[project]
name = "labby-voice"
version = "0.1.0"
description = "Voice-enabled Azure agent for Microsoft Teams"
requires-python = ">=3.12"
dependencies = [
    # Microsoft 365 Agents SDK
    "microsoft-agents-hosting-aiohttp",
    "microsoft-agents-hosting-teams",
    "microsoft-agents-authentication-msal",

    # Azure SDKs
    "azure-mgmt-resourcegraph",
    "azure-identity",

    # Voice Live API
    "websockets>=13.0",

    # Web server
    "aiohttp>=3.9",

    # Config
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bot*", "voice*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
